fixes and refactoring
