+++ Examples/Run/Fatras/Common/CMakeLists.txt
@@ -1,9 +1,9 @@
 add_library(
   ActsExamplesFatrasCommon SHARED
-  Fatras.cpp FatrasOptions.cpp)
+  src/FatrasCommon.cpp src/FatrasOptions.cpp)
 target_include_directories(
   ActsExamplesFatrasCommon
-  PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
+  PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>)
 target_link_libraries(
   ActsExamplesFatrasCommon
   PUBLIC
