+++ Examples/Run/Fatras/Common/src/FatrasCommon.cpp
@@ -6,8 +6,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "Fatras.hpp"
-
+#include "ActsExamples/Fatras/FatrasCommon.hpp"
+#include "ActsExamples/Options/FatrasOptions.hpp"
 #include "ActsExamples/Detector/IBaseDetector.hpp"
 #include "ActsExamples/Fatras/FatrasSimulation.hpp"
 #include "ActsExamples/Framework/RandomNumbers.hpp"
@@ -31,7 +31,6 @@
 
 #include <boost/program_options.hpp>
 
-#include "FatrasOptions.hpp"
 
 using namespace ActsExamples;
