+++ Examples/Run/Fatras/Common/src/FatrasOptions.cpp
@@ -6,9 +6,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "FatrasOptions.hpp"
-
 #include "Acts/Definitions/Units.hpp"
+#include "ActsExamples/Options/FatrasOptions.hpp"
 #include "ActsExamples/Fatras/FatrasSimulation.hpp"
 #include "ActsExamples/Utilities/Options.hpp"
