+++ Examples/Run/Fatras/DD4hep/DD4hepFatrasExample.cpp
@@ -7,8 +7,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #include "ActsExamples/DD4hepDetector/DD4hepDetector.hpp"
-
-#include "Fatras.hpp"
+#include "ActsExamples/Fatras/FatrasCommon.hpp"
 
 int main(int argc, char* argv[]) {
   return runFatras(argc, argv,
