+++ Examples/Run/Geant4/Common/src/Geant4Common.cpp
@@ -14,7 +14,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "Geant4.hpp"
 
 #include "Acts/Geometry/TrackingGeometry.hpp"
 #include "Acts/MagneticField/MagneticFieldProvider.hpp"
@@ -36,6 +35,7 @@
 #include "ActsExamples/Options/ParticleGunOptions.hpp"
 #include "ActsExamples/Simulation/CommonSimulation.hpp"
 #include "ActsExamples/Utilities/Paths.hpp"
+#include "ActsExamples/Geant4/Geant4Common.hpp"
 
 #include <memory>
 #include <string>
