+++ Examples/Run/Geant4/DD4hep/DD4hepMaterialRecording.cpp
@@ -11,11 +11,10 @@
 #include "ActsExamples/DDG4/DDG4DetectorConstruction.hpp"
 #include "ActsExamples/Options/CommonOptions.hpp"
 #include "ActsExamples/Options/ParticleGunOptions.hpp"
+#include "ActsExamples/Geant4/Geant4Common.hpp"
+#include "ActsExamples/Options/Geant4Options.hpp"
 
 #include <boost/program_options.hpp>
-
-#include "Geant4.hpp"
-#include "Geant4Options.hpp"
 
 using namespace ActsExamples;
