+++ Examples/Run/Geant4/GdmlMaterialRecording.cpp
@@ -9,12 +9,10 @@
 #include "ActsExamples/Geant4/GdmlDetectorConstruction.hpp"
 #include "ActsExamples/Options/CommonOptions.hpp"
 #include "ActsExamples/Options/ParticleGunOptions.hpp"
+#include "ActsExamples/Geant4/Geant4Common.hpp"
+#include "ActsExamples/Options/Geant4Options.hpp"
 
 #include <boost/program_options.hpp>
-
-// Geant4 Common includes
-#include "Geant4.hpp"
-#include "Geant4Options.hpp"
 
 using namespace ActsExamples;
