+++ Examples/Run/Geant4/HepMC/CMakeLists.txt
@@ -10,6 +10,8 @@ target_link_libraries(
     ActsExamplesFramework
     ActsExamplesGeant4
     ActsExamplesIoHepMC3
+    ActsExamplesHepMC3Options
+    ActsExamplesGeant4Common
     Boost::program_options)
     
 install(
