+++ Examples/Run/Geant4/HepMC/EventRecordingExample.cpp
@@ -13,19 +13,19 @@
 #include "ActsExamples/Framework/Sequencer.hpp"
 #include "ActsExamples/Geant4HepMC/EventRecording.hpp"
 #include "ActsExamples/Geometry/CommonGeometry.hpp"
-#include "ActsExamples/Io/Csv/CsvOptionsReader.hpp"
+#include "ActsExamples/Options/CsvOptionsReader.hpp"
 #include "ActsExamples/Io/Csv/CsvParticleReader.hpp"
-#include "ActsExamples/Io/HepMC3/HepMC3Options.hpp"
+#include "ActsExamples/Options/HepMC3Options.hpp"
 #include "ActsExamples/Io/HepMC3/HepMC3Writer.hpp"
 #include "ActsExamples/Options/CommonOptions.hpp"
 #include "ActsExamples/Utilities/Paths.hpp"
+#include "ActsExamples/Options/Geant4Options.hpp"
 
 #include <fstream>
 #include <string>
 
 #include <boost/program_options.hpp>
 
-#include "Geant4Options.hpp"
 
 int main(int argc, char* argv[]) {
   // Declare the supported program options.
