+++ Examples/Run/HepMC3/HepMC3HistogramExample.cpp
@@ -7,7 +7,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #include "ActsExamples/HepMC/HepMCProcessExtractor.hpp"
-#include "ActsExamples/Io/HepMC3/HepMC3Options.hpp"
+#include "ActsExamples/Options/HepMC3Options.hpp"
 #include "ActsExamples/Io/HepMC3/HepMC3Reader.hpp"
 #include "ActsExamples/Io/NuclearInteractions/RootNuclearInteractionParametersWriter.hpp"
 #include "ActsExamples/Options/CommonOptions.hpp"
