+++ Examples/Run/HepMC3/Options/include/ActsExamples/Options/HepMC3Options.hpp
@@ -9,9 +9,8 @@
 #pragma once
 
 #include "ActsExamples/Utilities/OptionsFwd.hpp"
-
-#include "HepMC3Reader.hpp"
-#include "HepMC3Writer.hpp"
+#include "ActsExamples/Io/HepMC3/HepMC3Reader.hpp"
+#include "ActsExamples/Io/HepMC3/HepMC3Writer.hpp"
 
 namespace ActsExamples {
 namespace Options {
