+++ Examples/Run/HepMC3/Options/src/HepMC3Options.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "ActsExamples/Io/HepMC3/HepMC3Options.hpp"
+#include "ActsExamples/Options/HepMC3Options.hpp"
 
 #include "ActsExamples/Utilities/Options.hpp"
